requirements.txt
discord.py==2.4.0
aiosqlite==0.20.0
pyyaml==6.0.1
stripe==11.1.1
tzdata==2025.1